Fixed intellisense problems with sexy.global.functions.cpp by breaking it up into smaller cpp files.
